Move App and Routes into shared folder
